requirements.txt
absl-py
astor
cachetools
certifi
chardet
click
Flask
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
gunicorn
h5py
idna
importlib-metadata
itsdangerous
Jinja2
joblib
Keras-Applications
Keras-Preprocessing
Markdown
MarkupSafe
numpy
oauthlib
opt-einsum
pickle-mixin
Pillow
protobuf
pyasn1
pyasn1-modules
requests
requests-oauthlib
rsa
scikit-learn
scipy
six
sklearn
tensorboard
tensorflow
tensorflow-estimator
termcolor
threadpoolctl
urllib3
Werkzeug
wrapt
zipp
